+++ compiler-tests/src/test/data/diagnostic/createGraph/CreateGraph_GraphHasFactory.kt
@@ -1,0 +1,14 @@
+// RENDER_DIAGNOSTICS_FULL_TEXT
+
+@DependencyGraph
+interface ExampleGraphWithFactory {
+  @DependencyGraph.Factory
+  interface Factory {
+    fun create(): ExampleGraphWithFactory
+  }
+}
+
+fun example() {
+  val exampleGraph = createGraph<<!CREATE_GRAPH_ERROR!>ExampleGraphWithFactory<!>>()
+  val exampleGraph2: ExampleGraphWithFactory = <!CREATE_GRAPH_ERROR!>createGraph<!>()
+}