+++ compiler-tests/src/test/data/diagnostic/createGraph/CreateGraph_OkCase.kt
@@ -1,0 +1,17 @@
+@DependencyGraph
+interface ExampleGraph
+
+@DependencyGraph
+interface ExampleGraphWithFactory {
+  @DependencyGraph.Factory
+  interface Factory {
+    fun create(): ExampleGraphWithFactory
+  }
+}
+
+fun example() {
+  val exampleGraph = createGraph<ExampleGraph>()
+  val exampleGraphWithFactory = createGraphFactory<ExampleGraphWithFactory.Factory>().create()
+  val exampleGraph2: ExampleGraph = createGraph()
+  val exampleGraphFactory: ExampleGraphWithFactory.Factory = createGraphFactory()
+}