+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/MetroFirBuiltIns.kt
@@ -27,6 +27,21 @@ internal class MetroFirBuiltIns(
     }
   }
 
+  val createGraph by unsafeLazy {
+    session.symbolProvider
+      .getTopLevelFunctionSymbols(Symbols.FqNames.metroRuntimePackage, "createGraph".asName())
+      .first()
+  }
+
+  val createGraphFactory by unsafeLazy {
+    session.symbolProvider
+      .getTopLevelFunctionSymbols(
+        Symbols.FqNames.metroRuntimePackage,
+        "createGraphFactory".asName(),
+      )
+      .first()
+  }
+
   val injectedFunctionClassClassSymbol by unsafeLazy {
     session.symbolProvider.getClassLikeSymbolByClassId(Symbols.ClassIds.metroInjectedFunctionClass)
       as FirRegularClassSymbol